Multidim file format (#6)
* use omfileformatc-rs crate

* add serde and bump omfileformatc

* omfile json struct

* decoding with Clib working

* bump omfileformatc-rs

* integrate clib fixes

* deactivate unused code

* WIP wrapping C code for writer and working on encoder

* fix write_next_chunks

* more tests

* more tests, but need to be fixed...

* fix write_chunks test

* fix tests

* run tests sequentially

* Use c encoder (#7)

* bump omfileformatc

* update fileformatc agan

* wip: implement c encoder

* reorganize tests

* no extra module named tests

* wip adopt c code

* lifetime fix

* fix offset write test

* fix write_3d test

* more tests

* old writer new reader test

* cleanup

* make old reader work again correctly

* all tests are working

* cleanup

* cleanup

* fix benches

* cleanup prints

* run checks on all pull requests

* Test Matrix on Linux, Mac and Windows (#9)

* test matrix

* run on any PR

* no mmap advice on windows

* fix for different types on msvc

* compression as well

* for now only remove files before test

* Bump c code (#10)

* WIP

* more wip

* WIP: build basically working

* wip fix tests

* WIP buffered write

* wip remove json

* wip fix tests

* more tests

* all tests

* cleanup

* more cleanup

* error cleanup

* imports

* improve error handling

* rc for child backends

* remove arc

* reformat script and small bug fix in old writer log10

* streaming writes for old writer

* fix benchmarks

* cleanup imports

* bin folder

* reactivate tests

* u64 usize wrangling

* error cleanup

* mmap cleanup

* restructure library

* string conversion fix on different platforms

* various small fixes and changes because of underlying submodule (#11)

* bump submodule

* remove OmOffsetSize_t, but some things still don't align fully...

* binary asserts

* better type magix

* fix test

* remove test for debugging

* writer2 usize cleanup

* remove unneeded lifetime

* bump submodule

* deactivate flaky test

* omfileformatc: back to old version without debug stuff

* Arc for reader backend

* implement writer backend for File

* OmFileReaderBackend provides flexibility to use owned data and references

* do not propage all errors

* always try to read owned data first...

* store header and trailer in OmFileReader2 to keep them in scope

* only keep variable in scope

* bump fileformat and related fixes

* bump submodule

+++ src/core/dimensions.rs
@@ -1,4 +1,4 @@
-use crate::om::errors::OmFilesRsError;
+use crate::errors::OmFilesRsError;
 use crate::utils::divide_rounded_up;
 use std::ops::Range;
 
@@ -38,8 +38,8 @@ impl Dimensions {
     }
 
     // length of the chunk offset table in bytes
-    pub fn chunk_offset_length(&self) -> usize {
-        self.n_chunks() * std::mem::size_of::<usize>()
+    pub fn chunk_offset_length(&self) -> u64 {
+        (self.n_chunks() * std::mem::size_of::<u64>()) as u64
     }
 
     #[inline(always)]
